Currently graph is working fine for bitrate in pcapfiles. Need to fix Y- axis titles, ts-files, changeStream doesnt work well for IAT dev.

+++ Status/streaminfo.h
@@ -31,7 +31,7 @@ public:
     NetworkProtocol protocol;
     BitrateMode bitrateMode;
     quint8 tsPerIp;
-    quint64 iatDeviation;
+    quint16 iatDeviation;
     quint64 lastSecondBytes;
     quint64 lastDuration;
     TsErrors tsErrors;
